+++ src/Types/Inline/QueryResult/Venue.php
@@ -44,9 +44,9 @@ class Venue extends AbstractInlineQueryResult
         'title' => true,
         'address' => true,
         'foursquare_id' => true,
-        'thumb_url' => true,
-        'thumb_width' => true,
-        'thumb_height' => true,
+        'thumbnail_url' => true,
+        'thumbnail_width' => true,
+        'thumbnail_height' => true,
         'reply_markup' => InlineKeyboardMarkup::class,
         'input_message_content' => InputMessageContent::class,
     ];
@@ -84,21 +84,21 @@ class Venue extends AbstractInlineQueryResult
      *
      * @var string|null
      */
-    protected $thumbUrl;
+    protected $thumbnailUrl;
 
     /**
      * Optional. Thumbnail width
      *
      * @var int|null
      */
-    protected $thumbWidth;
+    protected $thumbnailWidth;
 
     /**
      * Optional. Thumbnail height
      *
      * @var int|null
      */
-    protected $thumbHeight;
+    protected $thumbnailHeight;
 
     /**
      * Optional. Foursquare identifier of the venue if known
@@ -115,9 +115,9 @@ class Venue extends AbstractInlineQueryResult
      * @param float $longitude
      * @param string $title
      * @param string $address
-     * @param string|null $thumbUrl
-     * @param int|null $thumbWidth
-     * @param int|null $thumbHeight
+     * @param string|null $thumbnailUrl
+     * @param int|null $thumbnailWidth
+     * @param int|null $thumbnailHeight
      * @param string|null $foursquareId
      * @param InlineKeyboardMarkup|null $inlineKeyboardMarkup
      * @param InputMessageContent|null $inputMessageContent
@@ -128,9 +128,9 @@ class Venue extends AbstractInlineQueryResult
         $longitude,
         $title,
         $address,
-        $thumbUrl = null,
-        $thumbWidth = null,
-        $thumbHeight = null,
+        $thumbnailUrl = null,
+        $thumbnailWidth = null,
+        $thumbnailHeight = null,
         $foursquareId = null,
         $inputMessageContent = null,
         $inlineKeyboardMarkup = null
@@ -140,9 +140,9 @@ class Venue extends AbstractInlineQueryResult
         $this->latitude = $latitude;
         $this->longitude = $longitude;
         $this->address = $address;
-        $this->thumbUrl = $thumbUrl;
-        $this->thumbWidth = $thumbWidth;
-        $this->thumbHeight = $thumbHeight;
+        $this->thumbnailUrl = $thumbnailUrl;
+        $this->thumbnailWidth = $thumbnailWidth;
+        $this->thumbnailHeight = $thumbnailHeight;
         $this->foursquareId = $foursquareId;
     }
 
@@ -221,54 +221,120 @@ class Venue extends AbstractInlineQueryResult
     /**
      * @return string|null
      */
-    public function getThumbUrl()
+    public function getThumbnailUrl()
     {
-        return $this->thumbUrl;
+        return $this->thumbnailUrl;
     }
 
     /**
+     * @param string|null $thumbnailUrl
+     *
+     * @return void
+     */
+    public function setThumbnailUrl($thumbnailUrl)
+    {
+        $this->thumbnailUrl = $thumbnailUrl;
+    }
+
+    /**
+     * @deprecated Use getThumbnailUrl
+     *
+     * @return string|null
+     */
+    public function getThumbUrl()
+    {
+        return $this->getThumbnailUrl();
+    }
+
+    /**
+     * @deprecated Use setThumbnailUrl
+     *
      * @param string|null $thumbUrl
      *
      * @return void
      */
     public function setThumbUrl($thumbUrl)
     {
-        $this->thumbUrl = $thumbUrl;
+        $this->setThumbnailUrl($thumbUrl);
     }
 
     /**
      * @return int|null
      */
-    public function getThumbWidth()
+    public function getThumbnailWidth()
     {
-        return $this->thumbWidth;
+        return $this->thumbnailWidth;
     }
 
     /**
+     * @param int|null $thumbnailWidth
+     *
+     * @return void
+     */
+    public function setThumbnailWidth($thumbnailWidth)
+    {
+        $this->thumbnailWidth = $thumbnailWidth;
+    }
+
+    /**
+     * @deprecated Use getThumbnailWidth
+     *
+     * @return int|null
+     */
+    public function getThumbWidth()
+    {
+        return $this->getThumbnailWidth();
+    }
+
+    /**
+     * @deprecated Use setThumbnailWidth
+     *
      * @param int|null $thumbWidth
      *
      * @return void
      */
     public function setThumbWidth($thumbWidth)
     {
-        $this->thumbWidth = $thumbWidth;
+        $this->setThumbnailWidth($thumbWidth);
     }
 
     /**
      * @return int|null
      */
-    public function getThumbHeight()
+    public function getThumbnailHeight()
     {
-        return $this->thumbHeight;
+        return $this->thumbnailHeight;
     }
 
     /**
+     * @param int|null $thumbnailHeight
+     *
+     * @return void
+     */
+    public function setThumbnailHeight($thumbnailHeight)
+    {
+        $this->thumbnailHeight = $thumbnailHeight;
+    }
+
+    /**
+     * @deprecated Use getThumbnailHeight
+     *
+     * @return int|null
+     */
+    public function getThumbHeight()
+    {
+        return $this->getThumbnailHeight();
+    }
+
+    /**
+     * @deprecated Use setThumbnailWidth
+     *
      * @param int|null $thumbHeight
      *
      * @return void
      */
     public function setThumbHeight($thumbHeight)
     {
-        $this->thumbHeight = $thumbHeight;
+        $this->setThumbnailHeight($thumbHeight);
     }
 }